test_requirements.txt
django-nose
coverage
django-coverage
django-jasmine
ipdb
flake8
fabric
factory_boy>=2.0.0
mock
selenium
